# Backend/src/api/openapi.yaml
openapi: 3.0.3
info:
  title: Verve Dental Operational Suite API
  version: 0.6.0
  description: |
    Full OpenAPI contract for the Verve Operational Suite backend (Modules 03–08).

    ## Authentication

    All clinic-scoped endpoints require a Bearer JWT access token obtained from
    `POST /api/v1/auth/login` (direct) or `POST /api/v1/auth/mfa/verify` (MFA flow).

    The refresh token is transmitted as an HttpOnly cookie (`refreshToken`) scoped to
    `/api/v1/auth`. Use `POST /api/v1/auth/refresh` to rotate the access token without
    re-entering credentials.

    ### Standard Login Flow
    1. `POST /auth/login` with `{ email, password }`
    2. On success → `{ requiresMfa: false, accessToken, expiresIn, user }`
    3. Use `accessToken` as `Authorization: Bearer <token>` on subsequent requests.

    ### MFA Verification Flow (existing enrollment)
    1. `POST /auth/login` → `{ requiresMfa: true, mfaToken, user }`
    2. `POST /auth/mfa/verify` with `{ mfaToken, code }` → `{ accessToken, expiresIn, user }`

    ### MFA Enrollment Flow (first login for privileged roles)
    1. `POST /auth/login` → `{ requiresMfaEnrollment: true, enrollmentToken, user }`
    2. `POST /auth/mfa/setup` (Bearer = enrollmentToken) → `{ otpUri, secret }`
    3. User scans QR code in authenticator app.
    4. `POST /auth/mfa/confirm` (Bearer = enrollmentToken) with `{ code }` → confirms enrollment.
    5. Return to standard MFA verification flow.

    ## Roles

    | Role | Description |
...

...

...

...

...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from POST /api/v1/auth/login or /auth/mfa/verify.

  # ── Parameters ──────────────────────────────────────────────────────────────
  parameters:
    clinicId:
      name: clinicId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: >
        Tenant UUID. Must match the authenticated user's `homeClinicId`, or the
        caller must be `owner_admin`.
...
  responses:
    Unauthorized:
      description: Missing or invalid Bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: UNAUTHORIZED
              message: Authentication required
    Forbidden:
      description: Authenticated but not authorised for this resource or clinic.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: FORBIDDEN
              message: Insufficient permissions
...
    ValidationError:
      description: Request body or query parameter validation failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
          example:
            error:
              code: VALIDATION_ERROR
              message: Request validation failed
              details:
                - field: email
                  message: Invalid email address
...
  schemas:
...
    ValidationDetail:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
          description: >
            Dot-separated field path (e.g. "email", "address.postcode"). "_" for
            root-level issues.
          example: email
        message:
          type: string
          example: Invalid email address
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: NOT_FOUND
            message:
              type: string
              example: Resource not found
            requestId:
              type: string
              format: uuid
              nullable: true
            details:
              type: array
              items:
                $ref: "#/components/schemas/ValidationDetail"
    ValidationErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details]
          properties:
            code:
              type: string
              enum: [VALIDATION_ERROR]
            message:
              type: string
              example: Request validation failed
            requestId:
              type: string
              format: uuid
              nullable: true
            details:
              type: array
              items:
                $ref: "#/components/schemas/ValidationDetail"

    # ── Pagination ──────────────────────────────────────────────────────────────
...
    RosterEntry:
      type: object
      required:
        - id
        - staffUserId
        - staffEmail
        - rosteredClinicId
        - rosteredClinicName
        - shiftStartAt
        - shiftEndAt
        - shiftType
        - status
        - createdByUserId
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        staffUserId:
          type: string
          format: uuid
        staffEmail:
          type: string
          format: email
        rosteredClinicId:
          type: string
          format: uuid
          description: The clinic where the shift physically occurs. May differ from homeClinicId for cross-location deployments.
        rosteredClinicName:
          type: string
        shiftStartAt:
          type: string
          format: date-time
        shiftEndAt:
          type: string
          format: date-time
        shiftType:
          type: string
          enum: [standard, overtime, on_call, training]
        status:
          type: string
          enum: [scheduled, confirmed, completed, cancelled]
        notes:
          type: string
          nullable: true
        createdByUserId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
paths:
...
  /api/v1/clinics/{clinicId}/roster/me:
    get:
      summary: List the authenticated user's own shifts
      operationId: getMyShifts
      tags: [Roster]
      security:
        - bearerAuth: []
      description: >
        Returns shifts for the currently authenticated user. Accessible to all
        roles. Supports the same `from` / `to` filters as the full roster list.
        Does NOT return a pagination envelope — returns the array directly.
      parameters:
        - $ref: "#/components/parameters/clinicId"
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: List of the authenticated user's roster entries.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/RosterEntry"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...